% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{as_draws}
\alias{as_draws}
\alias{as_draws_list}
\alias{as_draws_array}
\alias{as_draws_df}
\alias{as_draws_matrix}
\alias{as_draws_rvars}
\alias{as_draws.ermod}
\alias{as_draws_list.ermod}
\alias{as_draws_array.ermod}
\alias{as_draws_df.ermod}
\alias{as_draws_matrix.ermod}
\alias{as_draws_rvars.ermod}
\title{Transform to \code{draws} objects}
\usage{
as_draws(x, ...)

as_draws_list(x, ...)

as_draws_array(x, ...)

as_draws_df(x, ...)

as_draws_matrix(x, ...)

as_draws_rvars(x, ...)

\method{as_draws}{ermod}(x, ...)

\method{as_draws_list}{ermod}(x, ...)

\method{as_draws_array}{ermod}(x, ...)

\method{as_draws_df}{ermod}(x, ...)

\method{as_draws_matrix}{ermod}(x, ...)

\method{as_draws_rvars}{ermod}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ermod}}

\item{...}{Arguments passed to individual methods (if applicable).}
}
\value{
A draws object from the \code{posterior} package.
}
\description{
See \code{\link[posterior:draws]{posterior::as_draws()}} for details.
}
