% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersim-methods.R
\name{calc_ersim_med_qi}
\alias{calc_ersim_med_qi}
\title{Calculate median and quantile intervals from ersim object}
\usage{
calc_ersim_med_qi(x, qi_width = 0.95)
}
\arguments{
\item{x}{An object of class \code{ersim} or \code{ersim_marg}}

\item{qi_width}{Width of the quantile interval}
}
\value{
An object of class \code{ersim_med_qi} or \code{ersim_marg_med_qi}
}
\description{
This is useful when you performed simulation with output_type = "draws" and
want to calculate median and quantile intervals without re-simulating.
}
