% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{d_sim_emax}
\alias{d_sim_emax}
\title{Sample simulated data for Emax exposure-response models with covariates.}
\format{
A data frame with columns:
\describe{
\item{dose}{Nominal dose, units not specified}
\item{exposure}{Exposure value, units and metric not specified}
\item{response_1}{Continuous response value (units not specified)}
\item{response_2}{Binary response value (group labels not specified)}
\item{cnt_a}{Continuous valued covariate}
\item{cnt_b}{Continuous valued covariate}
\item{cnt_c}{Continuous valued covariate}
\item{bin_d}{Binary valued covariate}
\item{bin_e}{Binary valued covariate}
}
}
\usage{
d_sim_emax
}
\description{
Sample simulated data for Emax exposure-response models with covariates.
}
\details{
This simulated dataset is entirely synthetic. It is a generic data set that can be used
to illustrate Emax modeling. It contains variables corresponding to dose and exposure,
and includes both a continuous response variable and a binary response variable. Three
continuous valued covariates are included, along with two binary covariates.

You can find the data generating code in the package source code,
under \code{data-raw/d_sim_emax.R}.
}
\examples{
d_sim_emax
}
\keyword{datasets}
