% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{d_sim_emax_3cov}
\alias{d_sim_emax_3cov}
\alias{d_sim_emax_bin_3cov}
\title{Sample simulated data for Emax exposure-response models with covariates.}
\format{
A data frame with columns:
\describe{
\item{dose}{Nominal dose, units specified}
\item{exposure}{Exposure value, units and metric not specified}
\item{cov_a}{Continuous valued covariate}
\item{cov_b}{Continuous valued covariate}
\item{cov_c}{Continuous valued covariate}
\item{response}{Response value (continuous for \code{d_sim_emax_3cov}, binary for \code{d_sim_emax_bin_3cov})}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 6 columns.
}
\usage{
d_sim_emax_3cov

d_sim_emax_bin_3cov
}
\description{
Sample simulated data for Emax exposure-response models with covariates.
}
\details{
This simulated dataset is...

You can find the data generating code in the package source code,
under \code{data-raw/d_sim_emax_3cov.R}.
}
\examples{
d_sim_emax_3cov
d_sim_emax_bin_3cov
}
\keyword{datasets}
