% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{d_sim_emax_cov}
\alias{d_sim_emax_cov}
\title{Sample simulated data for Emax exposure-response models with one covariate.}
\format{
A data frame with columns:
\describe{
\item{GP}{Group}
\item{c.is}{Concentration}
\item{eps}{Residual error}
\item{Emax.i}{Emax value}
\item{y.is}{Response value}
\item{Conc}{Concentration}
\item{Y}{Responbse}
}
}
\usage{
d_sim_emax_cov
}
\description{
Sample simulated data for Emax exposure-response models with one covariate.
}
\details{
This simulated dataset is...

You can find the data generating code in the package source code,
under \code{data-raw/d_sim_emax.R}.
}
\examples{
d_sim_emax_cov
}
\keyword{datasets}
