% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{d_sim_emax_nocov}
\alias{d_sim_emax_nocov}
\title{Sample simulated data for Emax exposure-response models with no covariates.}
\format{
A data frame with columns:
\describe{
\item{Conc}{Concentration}
\item{Y}{Response}
}
}
\usage{
d_sim_emax_nocov
}
\description{
Sample simulated data for Emax exposure-response models with no covariates.
}
\details{
This simulated dataset is...

You can find the data generating code in the package source code,
under \code{data-raw/d_sim_emax.R}.
}
\examples{
d_sim_emax_nocov
}
\keyword{datasets}
