% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyy.R
\docType{data}
\name{d_sim_lin}
\alias{d_sim_lin}
\title{Sample simulated data for exposure-response with continuous endpoint
using linear model.}
\format{
A data frame with columns:
\describe{
\item{ID}{Subject ID}
\item{AUCss}{Steady-state area under the curve}
\item{Cmaxss}{Steady-state maximum (peak) concentration}
\item{BAGE}{Baseline age in years}
\item{SEX}{M or F}
\item{response}{Response}
}
}
\usage{
d_sim_lin
}
\description{
Sample simulated data for exposure-response with continuous endpoint
using linear model.
}
\details{
True model is defined as \code{0.5 * AUCss + 0.5 * BAGE + 5 * SEX}, with
variability added with standard deviation of 10.
You can find the data generating code in the package source code,
under \code{data-raw/d_sim_lin.R}.
}
\examples{
d_sim_lin
}
\keyword{datasets}
