% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{ermod_cov_sel_method}
\alias{ermod_cov_sel_method}
\alias{print.ermod_cov_sel}
\alias{plot.ermod_cov_sel}
\title{S3 methods for the classes \code{ermod_bin_cov_sel}}
\usage{
\method{print}{ermod_cov_sel}(x, digits = 2, ...)

\method{plot}{ermod_cov_sel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ermod_bin_cov_sel}}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments passed to functions}
}
\value{
No return value, called for print or plot side effects
}
\description{
S3 methods for the classes \code{ermod_bin_cov_sel}
}
