% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{ermod_exp_sel_method}
\alias{ermod_exp_sel_method}
\alias{print.ermod_exp_sel}
\alias{plot.ermod_exp_sel}
\title{S3 methods for the classes \code{ermod_exp_sel}}
\usage{
\method{print}{ermod_exp_sel}(x, digits = 2, ...)

\method{plot}{ermod_exp_sel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ermod_bin_exp_sel}}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments passed to functions}
}
\description{
S3 methods for the classes \code{ermod_exp_sel}
}
