% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersim-methods.R
\name{ersim_method}
\alias{ersim_method}
\alias{plot.ersim}
\alias{plot.ersim_med_qi}
\title{S3 methods for the classes \verb{ersim_*} and \verb{ersim_med_qi_*}}
\usage{
\method{plot}{ersim}(x, show_orig_data = FALSE, ...)

\method{plot}{ersim_med_qi}(x, show_orig_data = FALSE, ...)
}
\arguments{
\item{x}{An object of the classes \verb{ersim_*} or \verb{ersim_med_qi_*}}

\item{show_orig_data}{logical, whether to show the data points in the
model development dataset. Default is \code{FALSE}. Only support plotting
with data that was used in the model development. If you want to use
other data, consider adding geom_point() to the plot manually.}

\item{...}{Additional arguments passed to functions}
}
\value{
No return value, called for print or plot side effects
}
\description{
S3 methods for the classes \verb{ersim_*} and \verb{ersim_med_qi_*}
}
