% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{loo}
\alias{loo}
\alias{loo.ermod}
\alias{loo.ermod_emax}
\alias{loo.ermod_bin_emax}
\title{Efficient approximate leave-one-out cross-validation (LOO)}
\usage{
loo(x, ...)

\method{loo}{ermod}(x, ...)

\method{loo}{ermod_emax}(x, ...)

\method{loo}{ermod_bin_emax}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ermod}}

\item{...}{Additional arguments passed to \code{loo::loo()}}
}
\value{
An object of class \code{loo}
}
\description{
See \code{\link[loo:loo]{loo::loo()}} for details.
}
