% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{prior_summary.ermod}
\alias{prior_summary.ermod}
\title{Summarize the priors used for linear or linear logistic regression models}
\usage{
\method{prior_summary}{ermod}(object, ...)
}
\value{
An object of class \code{prior_summary.stanreg}
}
\description{
See \code{\link[rstanarm:prior_summary.stanreg]{rstanarm::prior_summary()}} for details.
}
