% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesERtools-package.R
\docType{package}
\name{BayesERtools-package}
\alias{BayesERtools}
\alias{BayesERtools-package}
\title{BayesERtools: Bayesian Exposure-Response Analysis Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Suite of tools that facilitate exposure-response analysis using Bayesian methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genentech.github.io/BayesERtools/}
  \item \url{https://genentech.github.io/BayesERbook/}
}

}
\author{
\strong{Maintainer}: Kenta Yoshida \email{yoshida.kenta.6@gmail.com} (\href{https://orcid.org/0000-0003-4967-3831}{ORCID})

Authors:
\itemize{
  \item François Mercier \email{francois.mercier@roche.com} (\href{https://orcid.org/0000-0002-5685-1408}{ORCID})
}

Other contributors:
\itemize{
  \item Genentech, Inc. [copyright holder]
}

}
\keyword{internal}
