% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_ermod_lin.R
\name{dev_ermod_bin_exp_sel}
\alias{dev_ermod_bin_exp_sel}
\alias{dev_ermod_lin_exp_sel}
\title{Exposure metrics selection for linear ER models}
\usage{
dev_ermod_bin_exp_sel(
  data,
  var_resp,
  var_exp_candidates,
  options_placebo_handling = list(),
  prior = rstanarm::default_prior_coef(stats::binomial()),
  prior_intercept = rstanarm::default_prior_intercept(stats::binomial()),
  verbosity_level = 1,
  chains = 4,
  iter = 2000
)

dev_ermod_lin_exp_sel(
  data,
  var_resp,
  var_exp_candidates,
  options_placebo_handling = list(),
  prior = rstanarm::default_prior_coef(stats::binomial()),
  prior_intercept = rstanarm::default_prior_intercept(stats::binomial()),
  prior_aux = rstanarm::exponential(autoscale = TRUE),
  verbosity_level = 1,
  chains = 4,
  iter = 2000
)
}
\arguments{
\item{data}{Input data for E-R analysis}

\item{var_resp}{Response variable name in character}

\item{var_exp_candidates}{Candidate exposure variable names
in character vector}

\item{options_placebo_handling}{List of for placebo handling.
Possible options include:
\itemize{
\item \code{include_placebo}: Logical, whether the placebo group should be
used when estimating the model. Default is observed data. Default
is \code{FALSE}.
\item \code{method}: Character, specifying the method used to detect placebo
group samples. Possible values include \code{"zero_exposure_as_placebo"}
(the default), in which rows with zero values for the exposure
variable are automatically assigned to a placebo group,
\code{"var_placebo"}, indicating that the user will supply the name of
a data column specifying the placebo group, and \code{"none"}, in which
case no placebo handling takes place.
\item \code{var_placebo}: Character, specifying the name of a column in the
data set that indicates which rows belong to the placebo group. This
column is interpreted as a logical vector. This value is ignored
unless \code{method = "var_placebo"}.
}}

\item{prior, prior_intercept, prior_aux}{See \code{\link[rstanarm:stan_glm]{rstanarm::stan_glm()}}}

\item{verbosity_level}{Verbosity level. 0: No output, 1: Display steps,
2: Display progress in each step, 3: Display MCMC sampling.}

\item{chains}{Number of chains for Stan.}

\item{iter}{Number of iterations for Stan.}
}
\value{
An object of class \code{ermod_bin_exp_sel}.or \code{ermod_lin_exp_sel}
}
\description{
This functions is used to develop an linear ER model with binary
and continuous endpoint, using various exposure metrics and
selecting the best one.
}
\examples{
\donttest{
data(d_sim_binom_cov_hgly2)

ermod_bin_exp_sel <-
  dev_ermod_bin_exp_sel(
    data = d_sim_binom_cov_hgly2,
    var_resp = "AEFLAG",
    var_exp_candidates = c("AUCss_1000", "Cmaxss", "Cminss")
  )

ermod_bin_exp_sel
}

\donttest{
data(d_sim_lin)

ermod_lin_exp_sel <- dev_ermod_lin_exp_sel(
  data = d_sim_lin,
  var_resp = "response",
  var_exp_candidates = c("AUCss", "Cmaxss"),
  options_placebo_handling = list(use_placebo = TRUE)
)

ermod_lin_exp_sel
}

}
