% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{extract_coef_exp_ci}
\alias{extract_coef_exp_ci}
\title{Extract credible interval of the exposure coefficient}
\usage{
extract_coef_exp_ci(x, ci_width = 0.95)
}
\arguments{
\item{x}{An object of class \code{ermod_bin} or \code{ermod_lin}}

\item{ci_width}{Width of the credible interval}
}
\value{
A named vector of length 2 with the lower and upper bounds of the
credible interval (.lower, .upper)
}
\description{
Extract credible interval of the exposure coefficient
}
