% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{extract_ermod}
\alias{extract_ermod}
\alias{extract_data.ermod}
\alias{extract_mod.ermod}
\alias{extract_var_resp.ermod}
\alias{extract_var_exposure.ermod}
\alias{extract_var_cov.ermod}
\alias{extract_exp_sel_list_model.ermod_exp_sel}
\alias{extract_exp_sel_comp.ermod_exp_sel}
\alias{extract_var_selected.ermod_cov_sel}
\title{Extract elements from an object of class \code{ermod_*}}
\usage{
\method{extract_data}{ermod}(x)

\method{extract_mod}{ermod}(x)

\method{extract_var_resp}{ermod}(x)

\method{extract_var_exposure}{ermod}(x)

\method{extract_var_cov}{ermod}(x)

\method{extract_exp_sel_list_model}{ermod_exp_sel}(x)

\method{extract_exp_sel_comp}{ermod_exp_sel}(x)

\method{extract_var_selected}{ermod_cov_sel}(x)
}
\arguments{
\item{x}{An object of class \code{ermod_*}}
}
\value{
\itemize{
\item \code{\link[=extract_data]{extract_data()}} extracts data used for the model fit.
\item \code{\link[=extract_mod]{extract_mod()}} extracts the model fit object.
\item \code{\link[=extract_var_resp]{extract_var_resp()}} extracts the response variable name
\item \code{\link[=extract_var_exposure]{extract_var_exposure()}} extracts the exposure metric name
\item \code{\link[=extract_var_cov]{extract_var_cov()}} extracts the covariates name
\item \code{\link[=extract_exp_sel_list_model]{extract_exp_sel_list_model()}} extracts the list of fitted models for
each exposure metrics.
\item \code{\link[=extract_exp_sel_comp]{extract_exp_sel_comp()}} extracts the comparison results of the exposure
metrics.
\item \code{\link[=extract_var_selected]{extract_var_selected()}} extracts the selected variables (both exposure
and covariates)in the final model after covariate selection.
}
}
\description{
Extract elements from an object of class \code{ermod_*}
}
\keyword{internal}
