% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersim-methods.R
\name{extract_ersim}
\alias{extract_ersim}
\alias{extract_data.ersim}
\alias{extract_data.ersim_med_qi}
\alias{extract_var_resp.ersim}
\alias{extract_var_resp.ersim_med_qi}
\alias{extract_var_exposure.ersim}
\alias{extract_var_exposure.ersim_med_qi}
\alias{extract_var_cov.ersim}
\alias{extract_var_cov.ersim_med_qi}
\title{Extract elements from objects of the classes \verb{ersim_*} and `ersim_med_qi_*``}
\usage{
\method{extract_data}{ersim}(x)

\method{extract_data}{ersim_med_qi}(x)

\method{extract_var_resp}{ersim}(x)

\method{extract_var_resp}{ersim_med_qi}(x)

\method{extract_var_exposure}{ersim}(x)

\method{extract_var_exposure}{ersim_med_qi}(x)

\method{extract_var_cov}{ersim}(x)

\method{extract_var_cov}{ersim_med_qi}(x)
}
\arguments{
\item{x}{An object of class \code{ersim_*}}
}
\value{
\itemize{
\item \code{\link[=extract_data]{extract_data()}} extracts data used for the model fit.
\item \code{\link[=extract_mod]{extract_mod()}} extracts the model fit object.
\item \code{\link[=extract_var_resp]{extract_var_resp()}} extracts the response variable name
\item \code{\link[=extract_var_exposure]{extract_var_exposure()}} extracts the exposure metric name
\item \code{\link[=extract_var_cov]{extract_var_cov()}} extracts the covariates name
\item \code{\link[=extract_exp_sel_list_model]{extract_exp_sel_list_model()}} extracts the list of fitted models for
each exposure metrics.
\item \code{\link[=extract_exp_sel_comp]{extract_exp_sel_comp()}} extracts the comparison results of the exposure
metrics.
\item \code{\link[=extract_var_selected]{extract_var_selected()}} extracts the selected variables (both exposure
and covariates)in the final model after covariate selection.
}
}
\description{
Extract elements from objects of the classes \verb{ersim_*} and `ersim_med_qi_*``
}
\keyword{internal}
