% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermod-methods.R
\name{prior_summary}
\alias{prior_summary}
\alias{prior_summary.ermod}
\title{Summarize the priors used for linear or linear logistic regression models}
\usage{
prior_summary(object, ...)

\method{prior_summary}{ermod}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ermod}}

\item{...}{Additional arguments passed to \code{rstanarm::prior_summary()}}
}
\value{
An object of class \code{prior_summary.stanreg}
}
\description{
See \code{\link[rstanarm:prior_summary.stanreg]{rstanarm::prior_summary()}} for details.
}
